Routes - Inputs - Outputs - ProductByCategory

+++ src/app/components/products/products.component.ts
@@ -1,5 +1,5 @@
-import { Component } from '@angular/core';
-import { retry, switchMap, zip } from 'rxjs';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
+import { switchMap, zip } from 'rxjs';
 import { Product, CreateProductDTO, UpdateProductDTO } from 'src/app/models/product.model';
 import { ProductService } from 'src/app/services/product.service';
 import { StoreService } from 'src/app/services/store.service';
@@ -11,50 +11,23 @@ import { StoreService } from 'src/app/services/store.service';
 })
 export class ProductsComponent {
 
+  @Input() products: Product[] = [];
+  @Output() loadMore = new EventEmitter();
+
   myShoppingCartProducts: Product[] = []
   total = 0
-
-  products: Product[] = [];
 
   fechaHoy = new Date()
 
   showProductDetail = false
 
-  productChosen: Product = {
-    id: '',
-    price: 0,
-    images: [],
-    title: '',
-    category: {
-      id: '',
-      name: '',
-    },
-    description: ''
-  };
-
-  limit = 5
-  offset = 0
+  productChosen: Product | null = null
 
   statusDetail: 'loading' | 'success' | 'error' | 'init' = 'init';
 
   constructor(private storeService: StoreService,
     private productService: ProductService) {
     this.myShoppingCartProducts = this.storeService.GetMyShoppingCartProducts()
-  }
-
-  ngOnInit(): void {
-    // this.productService.GetAllProducts().subscribe(
-    // this.productService.getsProductsByPage(10, 0).subscribe(
-    //   response => {
-    //     console.log(response)
-    //     this.products = response
-    //     this.offset += this.limit
-    //   },
-    //   error => {
-    //     console.log(error)
-    //   }
-    // )
-    this.loadMore()
   }
 
   onAddToShoppingCart(product: Product) {
@@ -101,40 +74,35 @@ export class ProductsComponent {
   }
 
   updateProduct() {
-    const changes: UpdateProductDTO = {
-      title: "Nuevo Titulo",
+    if (this.productChosen) {
+      const changes: UpdateProductDTO = {
+        title: "Nuevo Titulo",
+      }
+      const id = this.productChosen?.id
+      this.productService.update(id, changes)
+        .subscribe(data => {
+          console.log("Update: ", data)
+          const productIndex = this.products.findIndex(item => item.id === this.productChosen?.id)
+          this.products[productIndex] = data
+        })
     }
-    const id = this.productChosen.id
-    this.productService.update(id, changes)
-      .subscribe(data => {
-        console.log("Update: ", data)
-        const productIndex = this.products.findIndex(item => item.id === this.productChosen.id)
-        this.products[productIndex] = data
-      })
   }
 
   deleteProduct() {
-    const id = this.productChosen.id
-    this.productService.delete(id).subscribe(
-      () => {
-        const productIndex = this.products.findIndex(item => item.id === this.productChosen.id)
-        this.products.splice(productIndex, 1)
-        this.showProductDetail = false
-      }
-    )
-  }
-
-  loadMore() {
-    return this.productService.getsProductsByPage(this.limit, this.offset)
-      .pipe(retry(3))
-      .subscribe(
-        data => {
-          this.products = this.products.concat(data)
-          this.offset += this.limit
-        }, error => {
-          console.log(error)
+    if (this.productChosen) {
+      const id = this.productChosen.id
+      this.productService.delete(id).subscribe(
+        () => {
+          const productIndex = this.products.findIndex(item => item.id === this.productChosen?.id)
+          this.products.splice(productIndex, 1)
+          this.showProductDetail = false
         }
       )
+    }
+  }
+
+  onLoadMore() {
+    this.loadMore.emit()
   }
 
   //Esto hay que evitarlo ya que podriamos caer en el callback hell
